Refactor test classes and add DaemonPollingService together with a test
This refactors the test-specific classes into a test package. The
DaemonPollingService will be responsible for continuously polling the
RPC of the wallet to keep us updated with the chain. This may or may not
be complemented with a notify event in the future. As it is, notify
callbacks can sometimes be a little unreliable with some Bitcoin-derived
code bases.